Remove Gerrit contact info filing step
This feature is no longer supported by the version of Gerrit that we
run. The decision was made to drop the need for contact filing because
the openstack.org website has an api now to do member lookups and we
don't have to rely on CLA signing in Gerrit for technical election
eligibility.

This coupled with Gerrit dropping support for the feature means we just
don't do it anymore.

Change-Id: I28bee0b5a8e8ed53fedd2369de54504369f5032f