Add common file for governance structures
This adds a page discussing Board, TC, UC, PTLs, ATCs, AUC, and
SIGs which is linked into the different presentations.

It is very similar to https://governance.openstack.org/ from which most
of the content is cribbed.

This tries to provide a limited overview with links to getting more
information.

Change-Id: I6440e84034832b9ebe552eec83438bea6ec3449a
Story: 2001359
Task: 5919
Task: 5920
Task: 5921