Add organization-guide stub
This patch adds an initial stub version of the new organization
guide. The purpose of this guide is to be something an employee
charged with contributing to OpenStack can take to their employer.
Describing things they may need to take into consideration to
help make the engagement fruitful and rewarding for everyone
involved.

At the moment I've just named it the organization guide. The
could be a better name. This document is a work in progress.

This came out of discussions the First Contact SIG had at the
Vancouver Forum[1]. Originially this discussion started a few months
before[2]. Fungi's post to the mailinglist sums it up much better
then I can[3].

[1] https://etherpad.openstack.org/p/Reqs-for-Organisations-Contributing-to-OpenStack
[2] https://etherpad.openstack.org/p/Contributing_Organization_Guide
[3] http://lists.openstack.org/pipermail/openstack-sigs/2018-June/000410.html

Change-Id: I9a48879959786f7f3883eec861cd5ca62ec90dcf