Change Ordering
Before this patch, task tracking was after the git section which didn't
make sense because in a normal workflow you would figure out what bug
or new feature you are working on before fiddling with starting a new
patch.

This patch also moves IRC to be right after the communication section and
events to be further up in the more overview themed sections.

Change-Id: I6b12e1efbecabaf25ab7deee8f9fdccbf5198913