Specifically call out Non-Code Contribution
Its not the most clear to people looking at the guide
that there are other ways to contribute other than code.
There are sections on being an operator and a user, but there
are other types of non code contribution than those. This is
the first step in enumerating these other types of contribution.

This patch also adds a section on mentoring to the common directory.
In the future, it might be good to add it to the indices for other
sections of the guide- code& docs, operators, etc- in a follow up patch
but at this point the focus is populating the non-code contribution
activities list.

Story: 2001437
Task: 6148

Story: 2004131
Task: 27577

Change-Id: I7b64ea62572749c0fe097766d12d5655df79b7c0