Reinstate warnings-as-errors
There was a time when doc build warnings were ignored. During this dark
period, several translations were introduced that generated warnings.
These went unnoticed. Then, in a separate repository, the sphinx -W flag
(treat warnings as errors) was switched on. But still, nobody noticed,
because nobody made changes to this repository.

Then came the day when someone did [1][2]. And their patches wouldn't
build. And they were sad.

So a workaround [3] was instated whereby, for this repository,
warnings-as-errors was switched off.

Then the broken translations were fixed [4][5][6]. And there was much
rejoicing.

So now we can reinstate -W, which will cause future borked translations
to break the build, enforcing appropriate discipline therein.

[1] https://review.openstack.org/#/c/633939/
[2] https://review.openstack.org/#/c/634333/
[3] https://review.openstack.org/#/c/633968/
[4] https://review.openstack.org/634545
[5] https://review.openstack.org/634592
[6] https://review.openstack.org/634638

Change-Id: I5e47a9147a59b0bcbaef116d8e4b91da0af806c5