Make a couple of links translatable
Sphinx lets you do links like this:

 Lorem ipsum dolor sit amet, consectetur `adipiscing elit`_

 .. adipiscing elit: https://foo.com/bar/baz.html#blah

In the rendered version, the "adipiscing elit" appears as a hyperlink to
the foo.com address; and the .. line doesn't show up at all.

This makes translators sad.

They either translate "adipiscing elit" and lose the hyperlink
(generating a warning when the translated document is built), or leave
it untranslated and wind up with English (or in this example, doglatin)
phrases in the middle of their sentence.

As a practice, this kind of linking should only be used in a
translatable document if the link text is the actual title of the
target. When it's natural words in a sentence, the following form is
preferred:

 Lorem ipsum dolor sit amet, consectetur `adipiscing elit <https://foo.com/bar/baz.html#blah>`_

This renders the exact same way; but now the translator can translate
"adipiscing elit" and preserve both the link and the language of the
sentence.

This patch switches out two such links.

Change-Id: I28bb450dd5525fe916bc982b64f727cf83a76f2c