Governance Fixes
The information about how many of each kind of board member
there are was specific to a point in time. To make it more general and
not have to update it regularly we can instead mention the maximum
number of members per type.

We can instead ask how many of each type are serving currently as an
exercise question.

Change-Id: Ibe21756cb40dcaf23085ae763f7faf9a980659bc