Refer to commit message "footers" not "tags"
Since Git already has an object called a "tag" it's best not to
confuse people by referring to the lines at the end of a commit
message as "tags" too. Instead just call them "footers" but retain
an anchor with the old tags name for that section in case anyone has
hyperlinked it elsewhere.

Change-Id: I13117028949d3e3cf2b1d06f30158e082f904c05