Update the references to the OpenStack Foundation to its new name
This patch changes OpenStack Foundation to Open Infrastructure
Foundation following the recent name change. It also updates the reference
to the Foundation individual memebreship account including the screenshot.

The patch also updates the outdated reference to the OpenStack Summit which
is now called Open Infrastructure Summit and is scheduled to happen annually.

Change-Id: Ia7b2246980c57a7c7355829276b53cdcae0cded7
Signed-off-by: Ildiko Vancsa <ildiko.vancsa@gmail.com>